% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amltrain}
\alias{amltrain}
\title{AML training data}
\format{
A data frame with 306 rows (subjects) and 322 columns:
\describe{
\item{cryr}{duration of complete response in years}
\item{relapse.death}{censoring indicator: 1 = relapsed or died; 0 = alive at last follow=up}
\item{ENSG00000001561}{normalized expression for indicated transcript}
\item{ENSG00000005249}{normalized expression for indicated transcript}
\item{ENSG00000006757}{normalized expression for indicated transcript}
\item{ENSG00000007062}{normalized expression for indicated transcript}
\item{ENSG00000007968}{normalized expression for indicated transcript}
\item{ENSG00000008283}{normalized expression for indicated transcript}
\item{ENSG00000008405}{normalized expression for indicated transcript}
\item{ENSG00000008441}{normalized expression for indicated transcript}
\item{ENSG00000010295}{normalized expression for indicated transcript}
\item{ENSG00000011028}{normalized expression for indicated transcript}
\item{ENSG00000011198}{normalized expression for indicated transcript}
\item{ENSG00000012779}{normalized expression for indicated transcript}
\item{ENSG00000012817}{normalized expression for indicated transcript}
\item{ENSG00000013306}{normalized expression for indicated transcript}
\item{ENSG00000013725}{normalized expression for indicated transcript}
\item{ENSG00000018189}{normalized expression for indicated transcript}
\item{ENSG00000022267}{normalized expression for indicated transcript}
\item{ENSG00000023171}{normalized expression for indicated transcript}
\item{ENSG00000023909}{normalized expression for indicated transcript}
\item{ENSG00000029639}{normalized expression for indicated transcript}
\item{ENSG00000047634}{normalized expression for indicated transcript}
\item{ENSG00000049192}{normalized expression for indicated transcript}
\item{ENSG00000053524}{normalized expression for indicated transcript}
\item{ENSG00000058056}{normalized expression for indicated transcript}
\item{ENSG00000060138}{normalized expression for indicated transcript}
\item{ENSG00000061918}{normalized expression for indicated transcript}
\item{ENSG00000065809}{normalized expression for indicated transcript}
\item{ENSG00000065923}{normalized expression for indicated transcript}
\item{ENSG00000068489}{normalized expression for indicated transcript}
\item{ENSG00000069020}{normalized expression for indicated transcript}
\item{ENSG00000070404}{normalized expression for indicated transcript}
\item{ENSG00000071894}{normalized expression for indicated transcript}
\item{ENSG00000072422}{normalized expression for indicated transcript}
\item{ENSG00000073605}{normalized expression for indicated transcript}
\item{ENSG00000076555}{normalized expression for indicated transcript}
\item{ENSG00000080823}{normalized expression for indicated transcript}
\item{ENSG00000089723}{normalized expression for indicated transcript}
\item{ENSG00000090382}{normalized expression for indicated transcript}
\item{ENSG00000090975}{normalized expression for indicated transcript}
\item{ENSG00000100068}{normalized expression for indicated transcript}
\item{ENSG00000100077}{normalized expression for indicated transcript}
\item{ENSG00000100299}{normalized expression for indicated transcript}
\item{ENSG00000100376}{normalized expression for indicated transcript}
\item{ENSG00000100418}{normalized expression for indicated transcript}
\item{ENSG00000100448}{normalized expression for indicated transcript}
\item{ENSG00000100596}{normalized expression for indicated transcript}
\item{ENSG00000100916}{normalized expression for indicated transcript}
\item{ENSG00000102409}{normalized expression for indicated transcript}
\item{ENSG00000102760}{normalized expression for indicated transcript}
\item{ENSG00000104689}{normalized expression for indicated transcript}
\item{ENSG00000104946}{normalized expression for indicated transcript}
\item{ENSG00000105518}{normalized expression for indicated transcript}
\item{ENSG00000105808}{normalized expression for indicated transcript}
\item{ENSG00000106367}{normalized expression for indicated transcript}
\item{ENSG00000106526}{normalized expression for indicated transcript}
\item{ENSG00000106546}{normalized expression for indicated transcript}
\item{ENSG00000106780}{normalized expression for indicated transcript}
\item{ENSG00000107104}{normalized expression for indicated transcript}
\item{ENSG00000107742}{normalized expression for indicated transcript}
\item{ENSG00000107798}{normalized expression for indicated transcript}
\item{ENSG00000107816}{normalized expression for indicated transcript}
\item{ENSG00000107957}{normalized expression for indicated transcript}
\item{ENSG00000109674}{normalized expression for indicated transcript}
\item{ENSG00000110076}{normalized expression for indicated transcript}
\item{ENSG00000110237}{normalized expression for indicated transcript}
\item{ENSG00000110492}{normalized expression for indicated transcript}
\item{ENSG00000110799}{normalized expression for indicated transcript}
\item{ENSG00000111275}{normalized expression for indicated transcript}
\item{ENSG00000112773}{normalized expression for indicated transcript}
\item{ENSG00000113504}{normalized expression for indicated transcript}
\item{ENSG00000114268}{normalized expression for indicated transcript}
\item{ENSG00000114737}{normalized expression for indicated transcript}
\item{ENSG00000115183}{normalized expression for indicated transcript}
\item{ENSG00000115414}{normalized expression for indicated transcript}
\item{ENSG00000115457}{normalized expression for indicated transcript}
\item{ENSG00000115525}{normalized expression for indicated transcript}
\item{ENSG00000116574}{normalized expression for indicated transcript}
\item{ENSG00000117480}{normalized expression for indicated transcript}
\item{ENSG00000119280}{normalized expression for indicated transcript}
\item{ENSG00000120594}{normalized expression for indicated transcript}
\item{ENSG00000120675}{normalized expression for indicated transcript}
\item{ENSG00000120832}{normalized expression for indicated transcript}
\item{ENSG00000120913}{normalized expression for indicated transcript}
\item{ENSG00000121005}{normalized expression for indicated transcript}
\item{ENSG00000121039}{normalized expression for indicated transcript}
\item{ENSG00000121274}{normalized expression for indicated transcript}
\item{ENSG00000123080}{normalized expression for indicated transcript}
\item{ENSG00000123836}{normalized expression for indicated transcript}
\item{ENSG00000124019}{normalized expression for indicated transcript}
\item{ENSG00000124882}{normalized expression for indicated transcript}
\item{ENSG00000126822}{normalized expression for indicated transcript}
\item{ENSG00000127152}{normalized expression for indicated transcript}
\item{ENSG00000129824}{normalized expression for indicated transcript}
\item{ENSG00000130702}{normalized expression for indicated transcript}
\item{ENSG00000131188}{normalized expression for indicated transcript}
\item{ENSG00000131370}{normalized expression for indicated transcript}
\item{ENSG00000132122}{normalized expression for indicated transcript}
\item{ENSG00000132530}{normalized expression for indicated transcript}
\item{ENSG00000132819}{normalized expression for indicated transcript}
\item{ENSG00000132849}{normalized expression for indicated transcript}
\item{ENSG00000133401}{normalized expression for indicated transcript}
\item{ENSG00000133619}{normalized expression for indicated transcript}
\item{ENSG00000134531}{normalized expression for indicated transcript}
\item{ENSG00000134897}{normalized expression for indicated transcript}
\item{ENSG00000135074}{normalized expression for indicated transcript}
\item{ENSG00000135245}{normalized expression for indicated transcript}
\item{ENSG00000135272}{normalized expression for indicated transcript}
\item{ENSG00000135362}{normalized expression for indicated transcript}
\item{ENSG00000135363}{normalized expression for indicated transcript}
\item{ENSG00000135916}{normalized expression for indicated transcript}
\item{ENSG00000136026}{normalized expression for indicated transcript}
\item{ENSG00000136193}{normalized expression for indicated transcript}
\item{ENSG00000136231}{normalized expression for indicated transcript}
\item{ENSG00000136997}{normalized expression for indicated transcript}
\item{ENSG00000137193}{normalized expression for indicated transcript}
\item{ENSG00000137198}{normalized expression for indicated transcript}
\item{ENSG00000138722}{normalized expression for indicated transcript}
\item{ENSG00000139318}{normalized expression for indicated transcript}
\item{ENSG00000140287}{normalized expression for indicated transcript}
\item{ENSG00000144036}{normalized expression for indicated transcript}
\item{ENSG00000144647}{normalized expression for indicated transcript}
\item{ENSG00000144677}{normalized expression for indicated transcript}
\item{ENSG00000145476}{normalized expression for indicated transcript}
\item{ENSG00000145545}{normalized expression for indicated transcript}
\item{ENSG00000146243}{normalized expression for indicated transcript}
\item{ENSG00000146373}{normalized expression for indicated transcript}
\item{ENSG00000147044}{normalized expression for indicated transcript}
\item{ENSG00000147180}{normalized expression for indicated transcript}
\item{ENSG00000148444}{normalized expression for indicated transcript}
\item{ENSG00000148484}{normalized expression for indicated transcript}
\item{ENSG00000149131}{normalized expression for indicated transcript}
\item{ENSG00000150760}{normalized expression for indicated transcript}
\item{ENSG00000150782}{normalized expression for indicated transcript}
\item{ENSG00000151135}{normalized expression for indicated transcript}
\item{ENSG00000151208}{normalized expression for indicated transcript}
\item{ENSG00000151458}{normalized expression for indicated transcript}
\item{ENSG00000152409}{normalized expression for indicated transcript}
\item{ENSG00000152580}{normalized expression for indicated transcript}
\item{ENSG00000152767}{normalized expression for indicated transcript}
\item{ENSG00000152778}{normalized expression for indicated transcript}
\item{ENSG00000153563}{normalized expression for indicated transcript}
\item{ENSG00000154217}{normalized expression for indicated transcript}
\item{ENSG00000154743}{normalized expression for indicated transcript}
\item{ENSG00000154760}{normalized expression for indicated transcript}
\item{ENSG00000154874}{normalized expression for indicated transcript}
\item{ENSG00000156381}{normalized expression for indicated transcript}
\item{ENSG00000157107}{normalized expression for indicated transcript}
\item{ENSG00000157240}{normalized expression for indicated transcript}
\item{ENSG00000157873}{normalized expression for indicated transcript}
\item{ENSG00000157978}{normalized expression for indicated transcript}
\item{ENSG00000158691}{normalized expression for indicated transcript}
\item{ENSG00000159339}{normalized expression for indicated transcript}
\item{ENSG00000159403}{normalized expression for indicated transcript}
\item{ENSG00000159788}{normalized expression for indicated transcript}
\item{ENSG00000160685}{normalized expression for indicated transcript}
\item{ENSG00000160781}{normalized expression for indicated transcript}
\item{ENSG00000161509}{normalized expression for indicated transcript}
\item{ENSG00000162433}{normalized expression for indicated transcript}
\item{ENSG00000162614}{normalized expression for indicated transcript}
\item{ENSG00000162676}{normalized expression for indicated transcript}
\item{ENSG00000163412}{normalized expression for indicated transcript}
\item{ENSG00000163702}{normalized expression for indicated transcript}
\item{ENSG00000163814}{normalized expression for indicated transcript}
\item{ENSG00000164086}{normalized expression for indicated transcript}
\item{ENSG00000164172}{normalized expression for indicated transcript}
\item{ENSG00000164442}{normalized expression for indicated transcript}
\item{ENSG00000165272}{normalized expression for indicated transcript}
\item{ENSG00000166165}{normalized expression for indicated transcript}
\item{ENSG00000166435}{normalized expression for indicated transcript}
\item{ENSG00000166987}{normalized expression for indicated transcript}
\item{ENSG00000167291}{normalized expression for indicated transcript}
\item{ENSG00000167565}{normalized expression for indicated transcript}
\item{ENSG00000167851}{normalized expression for indicated transcript}
\item{ENSG00000168026}{normalized expression for indicated transcript}
\item{ENSG00000168209}{normalized expression for indicated transcript}
\item{ENSG00000168502}{normalized expression for indicated transcript}
\item{ENSG00000168939}{normalized expression for indicated transcript}
\item{ENSG00000169203}{normalized expression for indicated transcript}
\item{ENSG00000169247}{normalized expression for indicated transcript}
\item{ENSG00000169504}{normalized expression for indicated transcript}
\item{ENSG00000169860}{normalized expression for indicated transcript}
\item{ENSG00000169991}{normalized expression for indicated transcript}
\item{ENSG00000170035}{normalized expression for indicated transcript}
\item{ENSG00000170180}{normalized expression for indicated transcript}
\item{ENSG00000170456}{normalized expression for indicated transcript}
\item{ENSG00000170522}{normalized expression for indicated transcript}
\item{ENSG00000170909}{normalized expression for indicated transcript}
\item{ENSG00000171121}{normalized expression for indicated transcript}
\item{ENSG00000171222}{normalized expression for indicated transcript}
\item{ENSG00000171476}{normalized expression for indicated transcript}
\item{ENSG00000171813}{normalized expression for indicated transcript}
\item{ENSG00000171962}{normalized expression for indicated transcript}
\item{ENSG00000172197}{normalized expression for indicated transcript}
\item{ENSG00000172236}{normalized expression for indicated transcript}
\item{ENSG00000173083}{normalized expression for indicated transcript}
\item{ENSG00000173530}{normalized expression for indicated transcript}
\item{ENSG00000173926}{normalized expression for indicated transcript}
\item{ENSG00000174059}{normalized expression for indicated transcript}
\item{ENSG00000174080}{normalized expression for indicated transcript}
\item{ENSG00000174130}{normalized expression for indicated transcript}
\item{ENSG00000174738}{normalized expression for indicated transcript}
\item{ENSG00000175265}{normalized expression for indicated transcript}
\item{ENSG00000175352}{normalized expression for indicated transcript}
\item{ENSG00000176597}{normalized expression for indicated transcript}
\item{ENSG00000179222}{normalized expression for indicated transcript}
\item{ENSG00000179630}{normalized expression for indicated transcript}
\item{ENSG00000179639}{normalized expression for indicated transcript}
\item{ENSG00000179820}{normalized expression for indicated transcript}
\item{ENSG00000180096}{normalized expression for indicated transcript}
\item{ENSG00000180596}{normalized expression for indicated transcript}
\item{ENSG00000180902}{normalized expression for indicated transcript}
\item{ENSG00000181104}{normalized expression for indicated transcript}
\item{ENSG00000182866}{normalized expression for indicated transcript}
\item{ENSG00000182871}{normalized expression for indicated transcript}
\item{ENSG00000183087}{normalized expression for indicated transcript}
\item{ENSG00000183091}{normalized expression for indicated transcript}
\item{ENSG00000184371}{normalized expression for indicated transcript}
\item{ENSG00000185129}{normalized expression for indicated transcript}
\item{ENSG00000185201}{normalized expression for indicated transcript}
\item{ENSG00000185245}{normalized expression for indicated transcript}
\item{ENSG00000185291}{normalized expression for indicated transcript}
\item{ENSG00000185304}{normalized expression for indicated transcript}
\item{ENSG00000185710}{normalized expression for indicated transcript}
\item{ENSG00000185883}{normalized expression for indicated transcript}
\item{ENSG00000185986}{normalized expression for indicated transcript}
\item{ENSG00000186130}{normalized expression for indicated transcript}
\item{ENSG00000186854}{normalized expression for indicated transcript}
\item{ENSG00000187010}{normalized expression for indicated transcript}
\item{ENSG00000187627}{normalized expression for indicated transcript}
\item{ENSG00000187653}{normalized expression for indicated transcript}
\item{ENSG00000187837}{normalized expression for indicated transcript}
\item{ENSG00000188002}{normalized expression for indicated transcript}
\item{ENSG00000188107}{normalized expression for indicated transcript}
\item{ENSG00000188211}{normalized expression for indicated transcript}
\item{ENSG00000188636}{normalized expression for indicated transcript}
\item{ENSG00000188738}{normalized expression for indicated transcript}
\item{ENSG00000188856}{normalized expression for indicated transcript}
\item{ENSG00000189164}{normalized expression for indicated transcript}
\item{ENSG00000189223}{normalized expression for indicated transcript}
\item{ENSG00000196155}{normalized expression for indicated transcript}
\item{ENSG00000196189}{normalized expression for indicated transcript}
\item{ENSG00000196415}{normalized expression for indicated transcript}
\item{ENSG00000196565}{normalized expression for indicated transcript}
\item{ENSG00000197081}{normalized expression for indicated transcript}
\item{ENSG00000197121}{normalized expression for indicated transcript}
\item{ENSG00000197253}{normalized expression for indicated transcript}
\item{ENSG00000197256}{normalized expression for indicated transcript}
\item{ENSG00000197321}{normalized expression for indicated transcript}
\item{ENSG00000197561}{normalized expression for indicated transcript}
\item{ENSG00000197728}{normalized expression for indicated transcript}
\item{ENSG00000197860}{normalized expression for indicated transcript}
\item{ENSG00000197937}{normalized expression for indicated transcript}
\item{ENSG00000197951}{normalized expression for indicated transcript}
\item{ENSG00000198743}{normalized expression for indicated transcript}
\item{ENSG00000198838}{normalized expression for indicated transcript}
\item{ENSG00000199347}{normalized expression for indicated transcript}
\item{ENSG00000200243}{normalized expression for indicated transcript}
\item{ENSG00000201801}{normalized expression for indicated transcript}
\item{ENSG00000203872}{normalized expression for indicated transcript}
\item{ENSG00000204172}{normalized expression for indicated transcript}
\item{ENSG00000205571}{normalized expression for indicated transcript}
\item{ENSG00000205593}{normalized expression for indicated transcript}
\item{ENSG00000208772}{normalized expression for indicated transcript}
\item{ENSG00000213085}{normalized expression for indicated transcript}
\item{ENSG00000213261}{normalized expression for indicated transcript}
\item{ENSG00000213626}{normalized expression for indicated transcript}
\item{ENSG00000213722}{normalized expression for indicated transcript}
\item{ENSG00000213906}{normalized expression for indicated transcript}
\item{ENSG00000213967}{normalized expression for indicated transcript}
\item{ENSG00000214016}{normalized expression for indicated transcript}
\item{ENSG00000214425}{normalized expression for indicated transcript}
\item{ENSG00000216316}{normalized expression for indicated transcript}
\item{ENSG00000220008}{normalized expression for indicated transcript}
\item{ENSG00000223345}{normalized expression for indicated transcript}
\item{ENSG00000224080}{normalized expression for indicated transcript}
\item{ENSG00000225138}{normalized expression for indicated transcript}
\item{ENSG00000226471}{normalized expression for indicated transcript}
\item{ENSG00000227097}{normalized expression for indicated transcript}
\item{ENSG00000227191}{normalized expression for indicated transcript}
\item{ENSG00000227615}{normalized expression for indicated transcript}
\item{ENSG00000228049}{normalized expression for indicated transcript}
\item{ENSG00000229153}{normalized expression for indicated transcript}
\item{ENSG00000230076}{normalized expression for indicated transcript}
\item{ENSG00000231160}{normalized expression for indicated transcript}
\item{ENSG00000231721}{normalized expression for indicated transcript}
\item{ENSG00000233927}{normalized expression for indicated transcript}
\item{ENSG00000233974}{normalized expression for indicated transcript}
\item{ENSG00000234883}{normalized expression for indicated transcript}
\item{ENSG00000236876}{normalized expression for indicated transcript}
\item{ENSG00000237298}{normalized expression for indicated transcript}
\item{ENSG00000237892}{normalized expression for indicated transcript}
\item{ENSG00000238160}{normalized expression for indicated transcript}
\item{ENSG00000239437}{normalized expression for indicated transcript}
\item{ENSG00000241399}{normalized expression for indicated transcript}
\item{ENSG00000241489}{normalized expression for indicated transcript}
\item{ENSG00000241529}{normalized expression for indicated transcript}
\item{ENSG00000244405}{normalized expression for indicated transcript}
\item{ENSG00000247627}{normalized expression for indicated transcript}
\item{ENSG00000249592}{normalized expression for indicated transcript}
\item{ENSG00000250116}{normalized expression for indicated transcript}
\item{ENSG00000250251}{normalized expression for indicated transcript}
\item{ENSG00000251079}{normalized expression for indicated transcript}
\item{ENSG00000253210}{normalized expression for indicated transcript}
\item{ENSG00000253276}{normalized expression for indicated transcript}
\item{ENSG00000254415}{normalized expression for indicated transcript}
\item{ENSG00000259276}{normalized expression for indicated transcript}
\item{ENSG00000260727}{normalized expression for indicated transcript}
\item{ENSG00000261377}{normalized expression for indicated transcript}
\item{ENSG00000264885}{normalized expression for indicated transcript}
\item{ENSG00000264895}{normalized expression for indicated transcript}
\item{ENSG00000267136}{normalized expression for indicated transcript}
\item{ENSG00000267551}{normalized expression for indicated transcript}
\item{ENSG00000267702}{normalized expression for indicated transcript}
\item{ENSG00000268001}{normalized expression for indicated transcript}
\item{ENSG00000268573}{normalized expression for indicated transcript}
\item{ENSG00000270554}{normalized expression for indicated transcript}
\item{ENSG00000270562}{normalized expression for indicated transcript}
\item{ENSG00000271646}{normalized expression for indicated transcript}
\item{ENSG00000273018}{normalized expression for indicated transcript}
\item{ENSG00000273033}{normalized expression for indicated transcript}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC11068580/}
}
\usage{
amltrain
}
\description{
Duration of complete response for 306 cytogenetically normal AML patients and a subset of 320 transcript expression from RNA-sequencing.
}
\keyword{datasets}
