% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{hdcuremodels-package}
\alias{hdcuremodels}
\alias{hdcuremodels-package}
\title{hdcuremodels: High-Dimensional Cure Models}
\description{
Provides functions for fitting various penalized parametric and semi-parametric mixture cure models with different penalty functions, testing for a significant cure fraction, and testing for sufficient follow-up as described in Fu et al (2022)\doi{10.1002/sim.9513} and Archer et al (2024)\doi{10.1186/s13045-024-01553-6}. False discovery rate controlled variable selection is provided using model-X knock-offs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kelliejarcher/hdcuremodels}
  \item Report bugs at \url{https://github.com/kelliejarcher/hdcuremodels/issues}
}

}
\author{
\strong{Maintainer}: Kellie J. Archer \email{archer.43@osu.edu} (\href{https://orcid.org/0000-0003-1555-5781}{ORCID})

Authors:
\itemize{
  \item Han Fu
}

Other contributors:
\itemize{
  \item Tung Lam Nguyen \email{nguyenlamtung10@gmail.com} (Reviewed the package for ROpenSci) [reviewer]
}

}
\keyword{internal}
