% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.mixturecure}
\alias{nobs.mixturecure}
\title{Number of observations in mixturecure object}
\usage{
\method{nobs}{mixturecure}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mixturecure}.}

\item{...}{other arguments.}
}
\value{
number of subjects in the dataset.
}
\description{
Number of observations in fitted \code{mixturecure} object.
}
\examples{
library(survival)
withr::local_seed(1234)
temp <- generate_cure_data(n = 100, j = 10, n_true = 10, a = 1.8)
training <- temp$training
fit <- curegmifs(Surv(Time, Censor) ~ .,
  data = training, x_latency = training,
  model = "weibull", thresh = 1e-4, maxit = 2000,
  epsilon = 0.01, verbose = FALSE
)
nobs(fit)
}
\keyword{methods}
